requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
